requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
black==23.9.1
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
decorator==5.1.1
Django==4.2.5
django-extensions==3.2.3
exceptiongroup==1.1.3
executing==2.0.0
Faker==19.11.0
fontawesomepro==6.4.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.16.1
isort==5.13.2
jedi==0.19.1
jmespath==1.0.1
Markdown==3.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.10.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.8
zope.interface==6.1
